# shifter/shifter_platform/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shifter Platform API",
    "version": "v1",
    "description": "Authenticated HTTP/JSON API for Shifter platform integrations."
  },
  "paths": {
...
    "/api/v1/mission-control/upload/cancel/": {
      "post": {
        "operationId": "api_v1_mission_control_upload_cancel",
        "description": "Cancel a validated current upload and clear the session marker.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCancel"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UploadCancel"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadCancel"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:upload:write"
        ]
      }
    },
    "/api/v1/mission-control/upload/complete/": {
      "post": {
        "operationId": "api_v1_mission_control_upload_complete",
        "description": "Finalize a completed upload and clear the session upload marker.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadComplete"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UploadComplete"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadComplete"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCompleteResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:upload:write"
        ]
      }
    },
    "/api/v1/mission-control/upload/initiate/": {
      "post": {
        "operationId": "api_v1_mission_control_upload_initiate",
        "description": "Create an upload session and presigned destination for an agent.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitiate"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitiate"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitiate"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitiateResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:upload:write"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Schema-only component for the canonical platform API error envelope.\n\nMirrors :func:`shared.api.errors.api_exception_handler` /\n:func:`shared.api.errors.api_error_response` output. It describes the wire\ncontract; it never runs at request time.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "Schema-only description of the ``shared.api.errors`` envelope body.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Safe, user-facing error message."
          },
          "details": {
            "description": "Optional structured field-level detail (present on validation errors)."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation id echoed from the X-Request-ID request header, when present."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "SuccessResponse": {
        "type": "object",
        "description": "Generic ``{\"success\": true}`` response body.\n\nShared by the range lifecycle mutations (cancel/destroy/pause/resume),\nupload cancel, and credential delete — every Mission Control mutation\nendpoint that has no payload beyond a boolean acknowledgement.",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "UploadCancel": {
        "type": "object",
        "description": "Validate agent-upload cancel requests.",
        "properties": {
          "upload_token": {
            "type": "string"
          }
        },
        "required": [
          "upload_token"
        ]
      },
      "UploadComplete": {
        "type": "object",
        "description": "Validate agent-upload completion requests.",
        "properties": {
          "upload_token": {
            "type": "string",
            "default": ""
          }
        }
      },
      "UploadCompleteResponse": {
        "type": "object",
        "description": "Response body for ``UploadCompleteView.post``.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "agent_id": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "message",
          "success"
        ]
      },
      "UploadInitiate": {
        "type": "object",
        "description": "Validate agent-upload initiation requests.",
        "properties": {
          "name": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "file_size": {},
          "agent_type": {
            "type": "string",
            "default": "xdr"
          }
        }
      },
      "UploadInitiateResponse": {
        "type": "object",
        "description": "Response body for ``UploadInitiateView.post`` (``cms.services.initiate_upload``).\n\n``presigned_url`` is a short-lived, single-use S3 PUT URL — an existing\nresponse field, typed here for the schema but never given a real example.",
        "properties": {
          "presigned_url": {
            "type": "string"
          },
          "s3_key": {
            "type": "string"
          },
          "upload_token": {
            "type": "string"
          },
          "expected_os": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "expected_os",
          "presigned_url",
          "s3_key",
          "upload_token"
        ]
      },
...
    },
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shf",
        "description": "Platform API token with scopes from shared.api_tokens.scopes."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  }
}